@@ -1,52 +1,119 @@
-name: lit_eracy
-description: A new Flutter project.
+# File: frontend/pubspec.yaml
 
+name: ai_literacy_app
+description: AI-powered educational app for children aged 3-14 with gamification and parental controls
 publish_to: 'none'
-
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.3.0 <4.0.0'
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_web_plugins:
-    sdk: flutter
-  cupertino_icons: ^1.0.2
-  dartz: ^0.10.1
-  dio: ^5.3.0
-  flutter_bloc: ^9.1.1
+
+  # State management
+  flutter_bloc: ^8.1.3
   equatable: ^2.0.5
-  get_it: ^8.0.3
+
+  # Dependency injection
+  get_it: ^7.6.0
+  injectable: ^2.3.2
+
+  # HTTP client and API calls
+  dio: ^5.3.2
+  retrofit: ^4.0.3
+  json_annotation: ^4.8.1
+
+  # Local storage
+  sqflite: ^2.3.0
   hive: ^2.2.3
   hive_flutter: ^1.1.0
-  path_provider: ^2.0.11
-  flutter_svg: ^2.0.5
-  camera: ^0.11.2
-  image_picker: ^1.1.2
-  tflite_flutter: ^0.11.0
-  image: ^4.0.15
-  logger: ^2.6.1
-  provider: ^6.0.5
-  shared_preferences: ^2.2.0
-  google_fonts: ^6.2.1
-  sqflite: ^2.3.0
-  connectivity_plus: ^6.1.4
-  path: ^1.8.3
-  bloc: ^9.0.0
-  http: ^1.4.0
-  
+  shared_preferences: ^2.2.2
+  path_provider: ^2.1.1
+
+  # Connectivity and networking
+  connectivity_plus: ^5.0.2
+  internet_connection_checker: ^1.0.0+1
+
+  # Authentication and security
+  flutter_secure_storage: ^9.0.0
+  crypto: ^3.0.3
+
+  # UI and animations
+  cupertino_icons: ^1.0.2
+  lottie: ^2.7.0
+  flutter_svg: ^2.0.9
+  cached_network_image: ^3.3.0
+  shimmer: ^3.0.0
+  flutter_spinkit: ^5.2.0
+  confetti: ^0.7.0
+  animated_text_kit: ^4.2.2
+
+  # Charts and data visualization
+  fl_chart: ^0.65.0
+  syncfusion_flutter_charts: ^23.2.7
+
+  # Utilities
+  intl: ^0.18.1
+  uuid: ^4.2.1
+  logger: ^2.0.2+1
+  collection: ^1.17.2
+  dartz: ^0.10.1
+
+  # Device features
+  permission_handler: ^11.1.0
+  device_info_plus: ^9.1.1
+  package_info_plus: ^4.2.0
+
+  # Audio and multimedia
+  audioplayers: ^5.2.1
+  flutter_tts: ^3.8.5
+
+  # Icons and fonts
+  font_awesome_flutter: ^10.6.0
+  google_fonts: ^6.1.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^6.0.0
-  hive_generator: ^2.0.0
-  build_runner: ^2.3.3
+
+  # Code generation
+  build_runner: ^2.4.7
+  injectable_generator: ^2.4.1
+  retrofit_generator: ^8.0.6
+  json_serializable: ^6.7.1
+  hive_generator: ^2.0.1
+
+  # Testing
+  mocktail: ^1.0.2
+  bloc_test: ^9.1.5
+
+  # Linting
+  flutter_lints: ^3.0.0
 
 flutter:
   uses-material-design: true
+
   assets:
     - assets/images/
+    - assets/icons/
+    - assets/animations/
+    - assets/sounds/
     - assets/models/
+
+  fonts:
+    - family: ComicNeue
+      fonts:
+        - asset: fonts/ComicNeue-Regular.ttf
+        - asset: fonts/ComicNeue-Bold.ttf
+          weight: 700
+    - family: Fredoka
+      fonts:
+        - asset: fonts/Fredoka-Regular.ttf
+        - asset: fonts/Fredoka-Medium.ttf
+          weight: 500
+        - asset: fonts/Fredoka-SemiBold.ttf
+          weight: 600
+        - asset: fonts/Fredoka-Bold.ttf
+          weight: 700
