@@ -1,43 +1,46 @@
-// frontend/lib/domain/entities/user.dart
+// # File: frontend/lib/domain/entities/user.dart
 
-class User {
+import 'package:equatable/equatable.dart';
+
+class User extends Equatable {
   final int id;
   final String username;
   final String email;
-  final String? fullName;
+  final String fullName;
   final int? age;
-  final String token;
+  final int? grade;
+  final String role;
+  final String? parentEmail;
+  final String? avatarUrl;
+  final bool isActive;
+  final DateTime createdAt;
 
   const User({
     required this.id,
     required this.username,
     required this.email,
-    this.fullName,
+    required this.fullName,
     this.age,
-    required this.token,
+    this.grade,
+    required this.role,
+    this.parentEmail,
+    this.avatarUrl,
+    required this.isActive,
+    required this.createdAt,
   });
 
-  factory User.fromJson(Map<String, dynamic> json) {
-    return User(
-      id: json['user']['id'] as int,
-      username: json['user']['username'] as String,
-      email: json['user']['email'] as String,
-      fullName: json['user']['full_name'] as String?,
-      age: json['user']['age'] as int?,
-      token: json['access_token'] as String,
-    );
-  }
-
-  Map<String, dynamic> toJson() {
-    return {
-      'user': {
-        'id': id,
-        'username': username,
-        'email': email,
-        if (fullName != null) 'full_name': fullName,
-        if (age != null) 'age': age,
-      },
-      'access_token': token,
-    };
-  }
+  @override
+  List<Object?> get props => [
+        id,
+        username,
+        email,
+        fullName,
+        age,
+        grade,
+        role,
+        parentEmail,
+        avatarUrl,
+        isActive,
+        createdAt,
+      ];
 }
