@@ -1,25 +1,37 @@
+# File: backend/app/database.py
+
 from sqlalchemy import create_engine
+from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 import os
-from dotenv import load_dotenv
-from .base import Base  # Import Base from separate file
 
-load_dotenv()
+# Database URL - supports both SQLite for development and PostgreSQL for production
+DATABASE_URL = os.getenv(
+    "DATABASE_URL", 
+    "sqlite:///./ai_literacy_app.db"
+)
 
-SQLALCHEMY_DATABASE_URL = os.getenv("DATABASE_URL", "sqlite:///./ai_literacy.db")
+# Handle PostgreSQL URL format for production
+if DATABASE_URL.startswith("postgres://"):
+    DATABASE_URL = DATABASE_URL.replace("postgres://", "postgresql://", 1)
 
-if SQLALCHEMY_DATABASE_URL.startswith("sqlite"):
+# Create engine
+if DATABASE_URL.startswith("sqlite"):
     engine = create_engine(
-        SQLALCHEMY_DATABASE_URL, 
-        connect_args={"check_same_thread": False}
+        DATABASE_URL,
+        connect_args={"check_same_thread": False}  # Only for SQLite
     )
 else:
-    engine = create_engine(SQLALCHEMY_DATABASE_URL)
+    engine = create_engine(DATABASE_URL)
 
+# Create session factory
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 
+# Base class for models
+Base = declarative_base()
+
+# Dependency to get database session
 def get_db():
-    """Dependency to get database session"""
     db = SessionLocal()
     try:
         yield db
