@@ -1,12 +1,32 @@
-abstract class Failure {
+// # File: frontend/lib/core/error/failures.dart
+
+import 'package:equatable/equatable.dart';
+
+abstract class Failure extends Equatable {
   final String message;
-  Failure(this.message);
+
+  const Failure(this.message);
+
+  @override
+  List<Object> get props => [message];
 }
 
 class ServerFailure extends Failure {
-  ServerFailure(super.message);
+  const ServerFailure(String message) : super(message);
 }
 
 class CacheFailure extends Failure {
-  CacheFailure(super.message);
+  const CacheFailure(String message) : super(message);
+}
+
+class NetworkFailure extends Failure {
+  const NetworkFailure(String message) : super(message);
+}
+
+class AuthFailure extends Failure {
+  const AuthFailure(String message) : super(message);
+}
+
+class ValidationFailure extends Failure {
+  const ValidationFailure(String message) : super(message);
 }
