@@ -1,21 +1,70 @@
+// # File: frontend/lib/presentation/blocs/auth/auth_bloc.dart
+
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:lit_eracy/domain/usecases/login_usecase.dart';
-import 'package:lit_eracy/domain/models/user.dart';
+import 'package:equatable/equatable.dart';
+
+import '../../../domain/usecases/auth/login_usecase.dart';
+import '../../../domain/usecases/auth/register_usecase.dart';
+import '../../../domain/entities/user.dart';
 
 part 'auth_event.dart';
 part 'auth_state.dart';
 
 class AuthBloc extends Bloc<AuthEvent, AuthState> {
-  final LoginUseCase loginUseCase;
+  final LoginUseCase _loginUseCase;
+  final RegisterUseCase _registerUseCase;
 
-  AuthBloc(this.loginUseCase) : super(AuthInitial()) {
-    on<LoginRequested>((event, emit) async {
-      emit(AuthLoading());
-      final result = await loginUseCase.execute(event.username, event.password);
-      result.fold(
-        (failure) => emit(AuthFailure(failure.message)),
-        (user) => emit(AuthSuccess(user)),
-      );
-    });
+  AuthBloc(this._loginUseCase, this._registerUseCase) : super(AuthInitial()) {
+    on<AuthInitialEvent>(_onAuthInitial);
+    on<LoginRequested>(_onLoginRequested);
+    on<RegisterRequested>(_onRegisterRequested);
+    on<LogoutRequested>(_onLogoutRequested);
+  }
+
+  void _onAuthInitial(AuthInitialEvent event, Emitter<AuthState> emit) {
+    emit(AuthInitial());
+  }
+
+  Future<void> _onLoginRequested(
+    LoginRequested event,
+    Emitter<AuthState> emit,
+  ) async {
+    emit(AuthLoading());
+
+    final result = await _loginUseCase(
+      LoginParams(username: event.username, password: event.password),
+    );
+
+    result.fold(
+      (failure) => emit(AuthFailure(failure.message)),
+      (user) => emit(AuthSuccess(user)),
+    );
+  }
+
+  Future<void> _onRegisterRequested(
+    RegisterRequested event,
+    Emitter<AuthState> emit,
+  ) async {
+    emit(AuthLoading());
+
+    final result = await _registerUseCase(
+      RegisterParams(
+        username: event.username,
+        email: event.email,
+        password: event.password,
+        fullName: event.fullName,
+        age: event.age,
+        grade: event.grade,
+      ),
+    );
+
+    result.fold(
+      (failure) => emit(AuthFailure(failure.message)),
+      (user) => emit(AuthSuccess(user)),
+    );
+  }
+
+  void _onLogoutRequested(LogoutRequested event, Emitter<AuthState> emit) {
+    emit(AuthInitial());
   }
 }
