@@ -1,6 +1,13 @@
+// # File: frontend/lib/presentation/blocs/auth/auth_state.dart
+
 part of 'auth_bloc.dart';
 
-abstract class AuthState {}
+abstract class AuthState extends Equatable {
+  const AuthState();
+
+  @override
+  List<Object?> get props => [];
+}
 
 class AuthInitial extends AuthState {}
 
@@ -8,10 +15,18 @@ class AuthLoading extends AuthState {}
 
 class AuthSuccess extends AuthState {
   final User user;
-  AuthSuccess(this.user);
+
+  const AuthSuccess(this.user);
+
+  @override
+  List<Object> get props => [user];
 }
 
 class AuthFailure extends AuthState {
   final String message;
-  AuthFailure(this.message);
+
+  const AuthFailure(this.message);
+
+  @override
+  List<Object> get props => [message];
 }
